fix: address high priority logic and network bugs (#186)
- Rethrow initialization errors to let Riverpod trigger the UI retry flow. 
- Add caching to course streams to stop redundant network fetches. 
- Fix permanent forum loading spinners and Django timer format crashes.

+++ packages/exams/lib/repositories/exam_repository.dart
@@ -1123,7 +1123,7 @@ class OnlineExamRepository implements ExamRepository {
     if (parts.length != 3) return 3600;
     return int.parse(parts[0]) * 3600 +
         int.parse(parts[1]) * 60 +
-        int.parse(parts[2]);
+        double.parse(parts[2]).toInt();
   }
 
   void dispose() {
